+++ src/main/ml-modules/root/ds/lux/advancedSearchConfig.mjs
@@ -1,10 +1,9 @@
 import { handleRequest } from '../../lib/securityLib.mjs';
 import { getAdvancedSearchConfig } from '../../config/advancedSearchConfig.mjs';
 
-const unitName = external.unitName;
 const response = handleRequest(function () {
   return getAdvancedSearchConfig();
-}, unitName);
+});
 
 response;
 export default response;