+++ src/main/ml-modules/root/ds/lux/autoComplete.mjs
@@ -1,7 +1,6 @@
 import { getMatches } from '../../lib/autoComple.mjs';
 import { handleRequest } from '../../lib/securityLib.mjs';
 
-const unitName = external.unitName;
 const text = external.text;
 const context = external.context;
 const fullyHonorContext = external.fullyHonorContext;
@@ -25,7 +24,7 @@ const response = handleRequest(function () {
     previouslyFiltered,
     timeoutInMilliseconds
   );
-}, unitName);
+});
 
 response;
 export default response;