+++ src/main/ml-modules/root/ds/lux/relatedList.mjs
@@ -1,7 +1,6 @@
 import { handleRequest } from '../../lib/securityLib.mjs';
 import { getRelatedList } from '../../lib/relatedListsLib.mjs';
 
-const unitName = external.unitName;
 const searchScopeName = external.scope;
 const relatedListName = external.name;
 const uri = external.uri;
@@ -20,7 +19,7 @@ const response = handleRequest(function () {
     filterResults,
     relationshipsPerRelation,
   });
-}, unitName);
+});
 
 response;
 export default response;