+++ src/test/ml-modules/root/test/suites/securityLibTests/0100 handleRequest.mjs
@@ -1,7 +1,6 @@
 import { testHelperProxy } from '/test/test-helper.mjs';
 import { executeScenario } from '/test/unitTestUtils.mjs';
-import { EndpointConfig } from '/lib/EndpointConfig.mjs';
-import { TENANT_OWNER, handleRequestForUnitTesting } from '/lib/securityLib.mjs';
+import { handleRequestForUnitTesting } from '/lib/securityLib.mjs';
 import { FOO_URI } from '/test/unitTestConstants.mjs';
 
 const LIB = '0100 handleRequest.mjs';
@@ -26,20 +25,9 @@ assertions.push(
 
 const scenarios = [
   {
-    name: 'Request against the tenant owner with a non-amp-as-admin endpoint',
+    name: 'Request returns the expected value',
     input: {
       function: returnBar,
-      unitName: TENANT_OWNER,
-      endpointConfig: { ampAsAdmin: false },
-    },
-    expected: { error: false, value: returnBar() },
-  },
-  {
-    name: 'Request with an undefined unit name defaults to the tenant owner',
-    input: {
-      function: returnBar,
-      unitName: undefined,
-      endpointConfig: { ampAsAdmin: false },
     },
     expected: { error: false, value: returnBar() },
   },
@@ -47,8 +35,6 @@ const scenarios = [
     name: 'Request able to access a document',
     input: {
       function: canReadDoc,
-      unitName: TENANT_OWNER,
-      endpointConfig: { ampAsAdmin: false },
     },
     expected: { error: false, value: true },
   },
@@ -56,11 +42,7 @@ const scenarios = [
 
 for (const scenario of scenarios) {
   const zeroArityFun = () => {
-    return handleRequestForUnitTesting(
-      scenario.input.function,
-      scenario.input.unitName,
-      new EndpointConfig(scenario.input.endpointConfig),
-    );
+    return handleRequestForUnitTesting(scenario.input.function);
   };
 
   const scenarioResults = executeScenario(scenario, zeroArityFun);